tests: Split Dell testing
A TestDell class is created to get a generic setUp().
This class is instanciated by products like c6615.
This will ease adding new Dell servers types for mocking.
